trial of error handling, validation and error codes

+++ controllers/customers_controller.py
@@ -4,7 +4,7 @@ from flask import request
 from main import db 
 from models.customers import Customer 
 from schemas.customers_schema import customer_schema, customers_schema
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 #! the customers controller is used to control end points for creating, retrieving, updating and deleting customers
 
@@ -20,17 +20,25 @@ def get_customers():
     #access all customers on database
     customers_list = Customer.query.all()
     result = customers_schema.dump(customers_list)
-    return jsonify(result)
+    return jsonify(result), 200
+
+
+
 
 #integer id specifies customer_id
 @customers.route('/<int:id>', methods=['GET'])
 def get_customer(id):
     #access 1 customer on database by choosing customer_id
     #get query will retrieve primary_key (which is customer_id)
-    #id could be changed to specific customer id that i want to retrieve, say cutomer_id 1, i could enter '1' instead of id and it will work the same.
     customer_list = Customer.query.get(id)
+    if not customer_list:
+        return {"Error": "Customer not found."}, 418
+        # yes, i know that the error code shouldve been 404 not found, but who can go past 418 :P
     result = customer_schema.dump(customer_list)
-    return jsonify(result)
+    return jsonify(result), 200
+
+
+
 
 #create new customer
 @customers.route('/', methods=['POST'])
@@ -47,6 +55,9 @@ def create_customer():
     db.session.commit()
     return jsonify(customer_schema.dump(customer))
 
+
+
+
 #update exsisting customer
 @customers.route('/<int:id>', methods=['PUT'])
 def update_customer(id):
@@ -62,21 +73,20 @@ def update_customer(id):
     db.session.commit()
     return jsonify(customer_schema.dump(customer))
 
+
+
+
 #DELETE customer 
 #! only available to admin
 @customers.route('/<int:id>', methods=['DELETE'])
 @jwt_required()
 def delete_customer(id):
+    if get_jwt_identity() != 'admin':
+        return {"DENIED": "You do not have permission to delete, only an administrator can delete. If you are an administrator, check your token and try again."}, 403
+
     customer = Customer.query.get(id)
     if not customer:
         return {"Error":"Could not find that customer, please enter existing customer id in order to delete a customer."}
     db.session.delete(customer)
     db.session.commit()
-    return {'Success': 'Customer successfully removed from database. This change is permenant, to re add the customer, POST the details.'}
-
-    
-#GET retrieves data from server 
-#POST sends data to server to create new resource 
-#PUT updates exsisting resource
-#PATCH similar to put, modifies exsisting resource. put method will contain complete new version, patch will contain only specific changes to the resouce
-#DELETE deletes resource specified by uri
+    return {'Success': 'Customer successfully removed from database. This change is permenant, to re add the customer, POST the details.'}, 200
